Added support to return arbitrary types (and sizes) in RMC replies.
Added an optimization to use registers if the reply size is 1, 4 or 8, otherwise use memcpy.

This change is needed to allow runtime_init() to be an RMC and return server memory region information.

+++ src/config.h
@@ -12,6 +12,6 @@ static constexpr const unsigned MAX_RMC_PROG_LEN = 56;
 /* rmc arguments */
 static constexpr const unsigned MAX_RMC_ARG_LEN = 16;
 /* to store RMC reply results */
-static constexpr const unsigned MAX_RMC_REPLY_LEN = 16;
+static constexpr const size_t MAX_RMC_REPLY_LEN = 16;
 /* how many RMCs we are going to register */
 static constexpr uint8_t NUM_REG_RMC = 1;
